ios 설정
