+++ frontend/README.md
@@ -26,7 +26,7 @@ Flutter 기반 음성 송금 인터페이스 시스템입니다.
 ---
 
 ### ✅ 2. 폴더 구조
-
+ 
 ```
 lib/
 ├── core/                  # 공통 유틸리티 (API 설정, 시간 기록 등)
