+++ frontend/ios/Runner/Info.plist
@@ -2,6 +2,12 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
+<key>NSMicrophoneUsageDescription</key>
+<string>이 앱은 음성 인식을 위해 마이크 접근이 필요합니다.</string>
+
+<key>NSLocationWhenInUseUsageDescription</key>
+<string>이 앱은 현재 위치를 사용하기 위해 위치 접근이 필요합니다.</string>
+
 	<key>CFBundleDevelopmentRegion</key>
 	<string>$(DEVELOPMENT_LANGUAGE)</string>
 	<key>CFBundleDisplayName</key>
