+++ frontend/lib/domain/repositories/stt_interface.dart
