+++ frontend/lib/presentation/providers/nlu_provider.dart
@@ -1,0 +1,6 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:voicetransfer/presentation/viewmodels/nlu_viewmodel.dart';
+
+final nluViewModelProvider = ChangeNotifierProvider<NluViewModel>((ref) {
+  return NluViewModel();
+});
